% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomog_D.R
\name{Dinhom}
\alias{Dinhom}
\title{Inhomogeneous Nearest neighbour function}
\usage{
Dinhom(x, r, intensity)
}
\arguments{
\item{x}{point pattern}

\item{r}{vector of ranges}

\item{intensity}{vector of intensities at points}
}
\description{
The same as spatstat's Ginhom with correction = "border"
}
