% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomog_Dcross.R
\name{Dinhom_cross}
\alias{Dinhom_cross}
\title{Inhomogeneous cross-Nearest neighbour function}
\usage{
Dinhom_cross(x, r, intensity, i = 1, j = 2, v2 = FALSE)
}
\arguments{
\item{x}{point pattern}

\item{intensity}{A vector, giving at point (x,m) the intensity intensity_m(x)}

\item{i}{Type from which to look. Index number(s).}

\item{j}{Type(s) to which to look. Index number(s).}

\item{v2}{dev, dont use.}
}
\description{
Compute the inhomogeneous D_ij(r). (or G_ij as in spatstat)
}
\details{
intensity could also be an im-object.
Note the difference: The infimum needed for the estimator
will be computed from the image, not just data locations.
}
