% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixon.R
\name{dixon}
\alias{dixon}
\title{Dixon's 2-type contingency table tests}
\usage{
dixon(X, prepR = 0)
}
\arguments{
\item{X}{Bivariate point pattern}

\item{prepR}{Preparation range, eases nearest neighbour computation for huge patterns.}
}
\description{
Computes the Dixon's test of segregation for a bivariate pattern using
the nearest neighbour contingency table.
}
\details{
The tests are an improvement on the Pielou's test of segregation. The test is defined only for two-type spatial pattern.
}
\references{
Philip Dixon, "Testing spatial segregation using a nearest-neighbor contingency
table", Ecology, 75, p.1940-1948 (1994).
}
