% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_bandwidth_profile}
\alias{intensity_bandwidth_profile}
\title{Find the optimal smoothing for intensity estimation kernel width}
\usage{
intensity_bandwidth_profile(x, bw_vector, scale = TRUE, ...)
}
\arguments{
\item{x}{point pattern}

\item{bw_vector}{Vector of bandwidth values to optimize over}

\item{scale}{Scale before computing? Recommended for numerical stability.}

\item{...}{passed on to density.ppp}
}
\value{
Vector of losses, or if keep_min=TRUE, list with a vector of losses and
a vector of intensity values with the bw that gives the minimum loss.
}
\description{
Find the optimal smoothing for intensity estimation kernel width
}
\details{
Optimize bandwidth $h$ of the Gaussian kernel using loss function
\deqn{latex}{\sum 1/\hat\lambda_h(x_i) - |W|)^2}

The bandwidth corresponds to the standard deviation of the Gaussian density.

The pattern will be scale to approx 1 unit window for numerical stability.
}

