% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAR_hom.R
\name{ISAR_homog}
\alias{ISAR_homog}
\title{Individual species-area-interaction or ISAR function}
\usage{
ISAR_homog(x, i, r, CSR = TRUE)
}
\arguments{
\item{x}{Point pattern}

\item{i}{i species}

\item{r}{r vector}

\item{CSR}{Include CSR? (not very useful)}
}
\description{
Individual species-area-interaction or ISAR function
}
\details{
CSR mean everything is Poisson. Not usually a realistic model.
}
