% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAR_inh.R
\name{ISAR_inhom}
\alias{ISAR_inhom}
\title{Inhomogeneous ISAR}
\usage{
ISAR_inhom(x, i, r, intensity, lambda = intensity, ...)
}
\arguments{
\item{x}{Point pattern}

\item{i}{Target species}

\item{r}{r vector}

\item{intensity}{Intensity lambda_m(x) at each point (x,m)}

\item{lambda}{same as intensity}

\item{...}{}
}
\description{
Inhomogeneous ISAR
}
