% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPSNC.R
\name{kernVarGamma}
\alias{kernVarGamma}
\title{Variance-Gamma (Bessel) kernel with bandwidth omega and shape parameter nu.ker}
\usage{
kernVarGamma(r, omega, nu.ker)
}
\description{
Variance-Gamma (Bessel) kernel with bandwidth omega and shape parameter nu.ker
}
