% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{plot.int_profile_stack}
\alias{plot.int_profile_stack}
\title{plot intensity optimisation}
\usage{
\method{plot}{int_profile_stack}(x, i, ...)
}
\arguments{
\item{x}{output of intensity_optimal}

\item{i}{which types. Default is the first.}

\item{...}{passed on to plot.intensity_profile}
}
\description{
plot intensity optimisation
}

