% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{summary.int_profile_stack}
\alias{summary.int_profile_stack}
\title{Summary for intensity optimal stack}
\usage{
\method{summary}{int_profile_stack}(x, ...)
}
\description{
Summary for intensity optimal stack
}
