% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{density_ppplist}
\alias{density_ppplist}
\title{Compute intensity surface for each component of multitype pattern}
\usage{
density_ppplist(x, sigma, ...)
}
\arguments{
\item{x}{a multitype pattern list}

\item{...}{passed on to density.ppp}

\item{bw}{the bandwidths, vector.}
}
\description{
Compute intensity surface for each component of multitype pattern
}
\details{
Uses density.ppp. Does what density.ppplist does but this one can take different bandwidths
for different types.
}

