% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_optimal}
\alias{intensity_optimal}
\title{Estimate the intensity at points using optimal bandwidth and within a type}
\usage{
intensity_optimal(x, verb = FALSE, ...)
}
\arguments{
\item{x}{point pattern, ppp object}

\item{verb}{verbosity}

\item{...}{passed on to \code{\link{intensity_bandwidth_profile}}}

\item{bwv}{Vector of bandwidths to optimise over}
}
\value{
A list with: a vector of intensities, one per point; the profiles; the best bandwidths.
}
\description{
Estimate the intensity at points using optimal bandwidth and within a type
}
\details{
First, we split the pattern per type. Then we optimise the
intensity using \code{\link{intensity_bandwidth_profile}}. Then
we concatenate the estimates to a vector.

Caution: For a big pattern with many types this will take time.
}

