% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAR.R
\name{ISAR}
\alias{ISAR}
\title{Homogeneous ISAR}
\usage{
ISAR(x, i, r, intensity, CSR = FALSE)
}
\arguments{
\item{x}{Multitype point pattern}

\item{i}{Target species}

\item{r}{r vector}

\item{intensity}{Optional vector of intensities, see details.}

\item{CSR}{Include CSR? Only for homogeneous case.}
}
\description{
Homogeneous ISAR
}
\details{
This function computes the ISAR function. The default action
is to compute the homogeneous version, that assumes all component patterns
are homogeneous (intensity is constant).

For inhomogeneous computation the \code{intensity} can be given. It should be
a vector giving the type-wise intensity at each point, or a list of im-objects giving the
intensity fields per mark. See \code{density.ppp} or \code{density_pplist} how to generate one.

CSR is not usually a good hypothesis: It assumes everything is homogeneous Poisson and independent.
}
\examples{
data(lansing, package = "spatstat")
# inhomogeneous
i1 <- ISAR(lansing, i = "hickory")
l <- intensity_optimal(lansing)
i2 <- ISAR(lansing, i = "hickory", intensity = l$intensity)
plot(i1)
lines(i2)

}
\seealso{
\code{\link{intensity_optimal}}
}

