% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAR_inh.R
\name{ISAR_inhom}
\alias{ISAR_inhom}
\title{Inhomogeneous ISAR}
\usage{
ISAR_inhom(x, i, r, intensity)
}
\arguments{
\item{x}{Point pattern}

\item{i}{Target species}

\item{r}{r vector}

\item{Intensity}{Intensity lambda_m(x) at each point (x,m)}
}
\description{
Inhomogeneous ISAR
}

