% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mingling.R
\name{mingling_index}
\alias{mingling_index}
\title{Mingling Index}
\usage{
mingling_index(x, i, r, normalise = FALSE, correction = "border",
  version = 1, ntype = "geometric")
}
\arguments{
\item{x}{multitype point pattern}

\item{i}{optional target type. If omitted, compute for all.}

\item{r}{neighbourhood radius / other parameter}

\item{normalise}{if TRUE, divide by CSR value 1 - lam[i]/lam.}

\item{correction}{"border" or "none" (defaut)}

\item{version}{1 or 2}

\item{ntype}{"geometric" or "knn"}
}
\description{
Compute the ratio of own type and all type neighbours
}
\details{
correction = "border" is minus-sampling. Version 1 is 1 - E(same type)/E(any type),
 version 2 is 1 - E [ (same type)/(any type) ]
}
